Repositories done

+++ zlatik/src/main/java/com/example/zlatik/repository/ContainsArtistRepository.java
@@ -3,5 +3,5 @@ package com.example.zlatik.repository;
 import com.example.zlatik.entity.ContainsArtist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContainsArtistRepository extends JpaRepository<ContainsArtist, Integer> {
+public interface ContainsArtistRepository extends JpaRepository<ContainsArtist, Long> {
 }
