+++ zlatik/src/main/java/com/example/zlatik/repository/ContainsSongRepository.java
@@ -3,5 +3,5 @@ package com.example.zlatik.repository;
 import com.example.zlatik.entity.ContainsSong;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContainsSongRepository extends JpaRepository<ContainsSong, Integer> {
+public interface ContainsSongRepository extends JpaRepository<ContainsSong, Long> {
 }
